child-sa: Add support for security labels
We install the configured label on the policies and the negotiated one
on the SAs.  This is how it usually is configured with SELinux where the
policy/configuration has a generic label and the SAs will get the actual
label of the flows assigned (the latter matches the former, so flows
match the policies but will trigger an acquire if no matching SA exists).